Add shop page components: ShopHeader, Accordion, ProductGrid, PriceRangeSlider, and ShopFilters
